Renamed example page and component registration
* Renamed example page to react_helmet
* Changed route to example page
* Renamed component registration to ReactHelmetApp
* Additional refactoring
