+++ spec/dummy/app/views/pages/_header.erb
@@ -78,7 +78,7 @@
     <%= link_to "Generator function returns object with renderedHtml", rendered_html_path %>
   </li>
   <li>
-    <%= link_to "Generator function returns object with renderedHtml as another object", rendered_htmls_path %>
+    <%= link_to "Generator function returns object with renderedHtml as another object", react_helmet_path %>
   </li>
   <li>
     <%= link_to "Image Example", image_example_path %>
