--- spec/dummy/app/views/pages/rendered_htmls.erb
@@ -1,15 +1,0 @@
-<%= render "header" %>
-
-<% render_hash = react_component("RenderedHtmls", prerender: true, props: { hello: "world" }, trace: true) %>
-
-<% content_for :title do %>
-  <%= render_hash['title'] %>
-<% end %>
-
-<%= render_hash["RenderedHtmls"] %>
-
-<hr/>
-
-This page demonstrates a generator function that returns htmlResult as an object
-with HTML strings on the server side.
-Check out the page title.
